qcacld-3.0: Call undo acs as part of deinit adapter
As part of SSR, hdd stop adapter is called which in
turn calls undo acs which frees the sap channel lists.
Now if SAP was already on, and then SSR happens then this
channel list would be freed.
After SSR the SAP would again be restarted.
Now if the driver gets an LTE-coex event from the FW,
there would be no channel-list left in the ACS list, as
it would freed, so there would be no suitable channel
for SAP restart leading to SAP stuck on an unsafe
channel which impact LTE throughput.

Fix is to call undo acs as part of deinit adapter which
would ensure that interface down has happened.

Change-Id: I5bf7bfcc20ccc686792b6b1df5d42b03288f7766
CRs-Fixed: 2650261
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>