qcacld-3.0: Fill proper seg0 and seg1 values
In the API lim_extract_ap_capability the driver
depends on the value of seg0 and seg1 values
from the AP and blindly fills that and sends
that as part of vdev start params.
There can be some APs which do not fill the
seg0 and seg1 APs correctly and blindly
depending upon those values can lead to
wrong seg0 and seg1 thus leading to FW assert.

Fix is to fill the seg0 value as centre channel
of 80Mhz and seg1 as 0 in case of 80Mhz BW.
Since these are unique values and dependent
on channel number (primary) driver can fill
that directly rather than depending on the
AP for them.

Change-Id: I46f2a81b443e1d0d62eb039878add15310595c8c
CRs-Fixed: 2664405
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>