qcacmn: Radar found agile center frequency for 160/80P80 MHz
When a radar is detected by the Agile detector, the radar found center
frequency is same as the frequency configured to the Agile detector.

This is true for operation in 20/40/80MHz. When operating in 160MHz mode,
the Agile center frquency is the center frequency of the 160MHz channel
and when operating in restricted 80P80MHz channel the Agile detector is
configured with 5690MHz as center frequency 1 and 5775MHz as center
frequency 2. In these cases the radar found frequency should be based on
the segment id on which the radar is detected.

Modify the radar found found agile center frequency based on the
segment id.

While modifying the center frequency in the function to get the
bonding channels, it is to be noted that the API
dfs_get_bonding_channels_for_freq() is based on HK Agile DFS model.
That is, when the current channel is 80P80/160MHz mode, the agile
channel is 80MHz only. Whereas, for Pine, when the current channel
is 80P80MHz, the agile channel is 160MHz mode and when the current
channel is 160MHz, the agile channel might be 80P80MHz/165MHz channel or
160MHz channel.

Accommodate Pine ADFS considerations in
dfs_get_bonding_channels_for_freq().

Change-Id: I31710a73a79a52f29aa4229bc55e08b81311b8d6
CRs-Fixed: 2670679
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>